adds a Generation tab in preset settings, consolidating several options
