+++ randovania/games/cave_story/gui/preset_settings/__init__.py
@@ -10,7 +10,7 @@ def cs_preset_tabs(editor: PresetEditor, window_manager: WindowManager):
 
     from randovania.gui.preset_settings.trick_level_tab import PresetTrickLevel
     from randovania.gui.preset_settings.starting_area_tab import PresetStartingArea
-    from randovania.gui.preset_settings.logic_damage_tab import PresetLogicDamage
+    from randovania.games.cave_story.gui.preset_settings.cs_generation_tab import PresetCSGeneration
     from randovania.gui.preset_settings.location_pool_tab import PresetLocationPool
     from randovania.games.cave_story.gui.preset_settings.cs_item_pool_tab import CSPresetItemPool
     from randovania.games.cave_story.gui.preset_settings.cs_goal_tab import PresetCSObjective
@@ -19,7 +19,7 @@ def cs_preset_tabs(editor: PresetEditor, window_manager: WindowManager):
     return [
         PresetTrickLevel(editor, game_description, window_manager),
         PresetStartingArea(editor, game_description),
-        PresetLogicDamage(editor),
+        PresetCSGeneration(editor),
         PresetCSObjective(editor),
         PresetLocationPool(editor, game_description),
         CSPresetItemPool(editor),
