+++ randovania/games/dread/gui/preset_settings/__init__.py
@@ -11,7 +11,7 @@ def dread_preset_tabs(editor: PresetEditor, window_manager: WindowManager):
     from randovania.gui.preset_settings.trick_level_tab import PresetTrickLevel
     from randovania.gui.preset_settings.elevators_tab import PresetElevators
     from randovania.gui.preset_settings.starting_area_tab import PresetStartingArea
-    from randovania.gui.preset_settings.logic_damage_tab import PresetLogicDamage
+    from randovania.gui.preset_settings.generation_tab import PresetGeneration
     from randovania.gui.preset_settings.location_pool_tab import PresetLocationPool
     from randovania.gui.preset_settings.metroid_item_pool_tab import MetroidPresetItemPool
 
@@ -19,7 +19,7 @@ def dread_preset_tabs(editor: PresetEditor, window_manager: WindowManager):
         PresetTrickLevel(editor, game_description, window_manager),
         PresetElevators(editor, game_description),
         PresetStartingArea(editor, game_description),
-        PresetLogicDamage(editor),
+        PresetGeneration(editor),
         PresetLocationPool(editor, game_description),
         MetroidPresetItemPool(editor)
     ]
