+++ randovania/games/prime1/gui/preset_settings/__init__.py
@@ -12,18 +12,18 @@ def prime1_preset_tabs(editor: PresetEditor, window_manager: WindowManager):
     from randovania.gui.preset_settings.patcher_energy_tab import PresetPatcherEnergy
     from randovania.gui.preset_settings.elevators_tab import PresetElevators
     from randovania.gui.preset_settings.starting_area_tab import PresetStartingArea
-    from randovania.gui.preset_settings.logic_damage_tab import PresetLogicDamage
     from randovania.games.prime1.gui.preset_settings.prime_goal_tab import PresetPrimeGoal
     from randovania.games.prime1.gui.preset_settings.prime_hints_tab import PresetPrimeHints
     from randovania.games.prime1.gui.preset_settings.prime_patches_tab import PresetPrimePatches
     from randovania.gui.preset_settings.location_pool_tab import PresetLocationPool
     from randovania.gui.preset_settings.metroid_item_pool_tab import MetroidPresetItemPool
+    from randovania.games.prime1.gui.preset_settings.prime_generation_tab import PresetPrimeGeneration
     return [
         PresetTrickLevel(editor, game_description, window_manager),
         PresetPatcherEnergy(editor, game_enum),
         PresetElevators(editor, game_description),
         PresetStartingArea(editor, game_description),
-        PresetLogicDamage(editor),
+        PresetPrimeGeneration(editor),
         PresetPrimeGoal(editor),
         PresetPrimeHints(editor),
         PresetPrimePatches(editor),
