+++ randovania/games/prime1/gui/preset_settings/prime_goal_tab.py
@@ -15,7 +15,6 @@ class PresetPrimeGoal(PresetTab, Ui_PresetPrimeGoal):
 
         self.goal_layout.setAlignment(QtCore.Qt.AlignTop)
         self.slider.valueChanged.connect(self._on_slider_changed)
-        self.min_progression_spin.valueChanged.connect(self._on_spin_changed)
 
     @property
     def uses_patches_tab(self) -> bool:
@@ -28,13 +27,6 @@ class PresetPrimeGoal(PresetTab, Ui_PresetPrimeGoal):
                 LayoutArtifactMode(self.slider.value())
             )
 
-    def _on_spin_changed(self):
-        with self._editor as editor:
-            editor.set_configuration_field(
-                "artifact_minimum_progression",
-                self.min_progression_spin.value(),
-            )
-
     def _on_slider_changed(self):
         self.slider_label.setText(str(self.slider.value()))
         self._update_editor()
@@ -42,4 +34,3 @@ class PresetPrimeGoal(PresetTab, Ui_PresetPrimeGoal):
     def on_preset_changed(self, preset: Preset):
         artifacts = preset.configuration.artifact_target
         self.slider.setValue(artifacts.num_artifacts)
-        self.min_progression_spin.setValue(preset.configuration.artifact_minimum_progression)
