+++ randovania/games/prime2/gui/preset_settings/__init__.py
@@ -11,7 +11,7 @@ def prime2_preset_tabs(editor: PresetEditor, window_manager: WindowManager):
     from randovania.gui.preset_settings.patcher_energy_tab import PresetPatcherEnergy
     from randovania.gui.preset_settings.elevators_tab import PresetElevators
     from randovania.gui.preset_settings.starting_area_tab import PresetStartingArea
-    from randovania.gui.preset_settings.logic_damage_tab import PresetLogicDamage
+    from randovania.gui.preset_settings.generation_tab import PresetGeneration
     from randovania.games.prime2.gui.preset_settings.echoes_goal_tab import PresetEchoesGoal
     from randovania.games.prime2.gui.preset_settings.echoes_hints_tab import PresetEchoesHints
     from randovania.games.prime2.gui.preset_settings.echoes_translators_tab import PresetEchoesTranslators
@@ -24,7 +24,7 @@ def prime2_preset_tabs(editor: PresetEditor, window_manager: WindowManager):
         PresetPatcherEnergy(editor, game_enum),
         PresetElevators(editor, game_description),
         PresetStartingArea(editor, game_description),
-        PresetLogicDamage(editor),
+        PresetGeneration(editor),
         PresetEchoesGoal(editor),
         PresetEchoesHints(editor),
         PresetEchoesTranslators(editor),
