+++ randovania/gui/ui_files/preset_item_pool.ui
@@ -122,16 +122,6 @@
          <number>0</number>
         </property>
         <item>
-         <widget class="QCheckBox" name="multi_pickup_placement_check">
-          <property name="toolTip">
-           <string>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p&gt;Use the new version of the pickup placement logic, which is able to place multiple pickups at once to solve a requirement, like both Seeker Launcher and Dark Visor. Not tested yet.&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</string>
-          </property>
-          <property name="text">
-           <string>Use multi-pickup placement [Experimental]</string>
-          </property>
-         </widget>
-        </item>
-        <item>
          <widget class="QGroupBox" name="random_starting_box">
           <property name="sizePolicy">
            <sizepolicy hsizetype="Preferred" vsizetype="Preferred">
