+++ randovania/gui/ui_files/preset_location_pool.ui
@@ -34,32 +34,6 @@
      <number>2</number>
     </property>
     <item>
-     <widget class="QGroupBox" name="randomization_mode_group">
-      <property name="title">
-       <string>Randomization Mode</string>
-      </property>
-      <layout class="QVBoxLayout" name="verticalLayout_2">
-       <item>
-        <widget class="QCheckBox" name="check_major_minor">
-         <property name="text">
-          <string>Enable major/minor split</string>
-         </property>
-        </widget>
-       </item>
-       <item>
-        <widget class="QLabel" name="randomization_mode_label">
-         <property name="text">
-          <string>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p&gt;If this setting is enabled, major items (i.e., major upgrades, Energy Tanks, Dark Temple Keys, and Energy Transfer Modules) and minor items (i.e, expansions) will be shuffled separately.&lt;br/&gt;Major items in excess of the number of major locations will be placed in minor locations, and vice versa.&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</string>
-         </property>
-         <property name="wordWrap">
-          <bool>true</bool>
-         </property>
-        </widget>
-       </item>
-      </layout>
-     </widget>
-    </item>
-    <item>
      <widget class="QScrollArea" name="locations_scroll_area">
       <property name="widgetResizable">
        <bool>true</bool>
@@ -70,7 +44,7 @@
          <x>0</x>
          <y>0</y>
          <width>499</width>
-         <height>343</height>
+         <height>457</height>
         </rect>
        </property>
        <layout class="QVBoxLayout" name="locations_scroll_area_layout">
