+++ randovania/gui/ui_files/preset_prime_goal.ui
@@ -85,16 +85,6 @@
       </item>
      </layout>
     </item>
-    <item>
-     <widget class="QLabel" name="min_progression_label">
-      <property name="text">
-       <string>Only place artifacts after this many actions were performed by the generator.</string>
-      </property>
-     </widget>
-    </item>
-    <item>
-     <widget class="QSpinBox" name="min_progression_spin"/>
-    </item>
    </layout>
   </widget>
  </widget>
