+++ randovania/gui/ui_files/preset_trick_level.ui
@@ -49,8 +49,8 @@
         <rect>
          <x>0</x>
          <y>0</y>
-         <width>525</width>
-         <height>625</height>
+         <width>539</width>
+         <height>516</height>
         </rect>
        </property>
        <layout class="QVBoxLayout" name="trick_level_layout">
@@ -87,68 +87,6 @@
          </widget>
         </item>
         <item>
-         <layout class="QHBoxLayout" name="dangerous_layout">
-          <item>
-           <widget class="QLabel" name="dangerous_label">
-            <property name="text">
-             <string>Dangerous actions:</string>
-            </property>
-           </widget>
-          </item>
-          <item>
-           <widget class="QComboBox" name="dangerous_combo">
-            <item>
-             <property name="text">
-              <string>Randomly</string>
-             </property>
-            </item>
-            <item>
-             <property name="text">
-              <string>Last Resort</string>
-             </property>
-            </item>
-           </widget>
-          </item>
-         </layout>
-        </item>
-        <item>
-         <widget class="QLabel" name="dangerous_description">
-          <property name="text">
-           <string>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p&gt;A dangerous action is the act of moving past a lock without the appropriate items needed to head backwards, or doing an action that can only be done once.&lt;/p&gt;&lt;p&gt;&lt;span style=&quot; font-weight:600;&quot;&gt;Randomly&lt;/span&gt;: Dangerous actions might be required by logic.&lt;/p&gt;&lt;p&gt;&lt;span style=&quot; font-weight:600;&quot;&gt;Last Resort&lt;/span&gt;: Only allows dangerous actions to be required if no other option is available for progression.&lt;br/&gt;Warning: Due to how item placement works, certain locations will have progression extremely less often or even never.&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</string>
-          </property>
-          <property name="wordWrap">
-           <bool>true</bool>
-          </property>
-         </widget>
-        </item>
-        <item>
-         <widget class="Line" name="trick_level_line_2">
-          <property name="orientation">
-           <enum>Qt::Horizontal</enum>
-          </property>
-         </widget>
-        </item>
-        <item>
-         <widget class="QCheckBox" name="trick_level_minimal_logic_check">
-          <property name="text">
-           <string>Use Minimal Logic</string>
-          </property>
-         </widget>
-        </item>
-        <item>
-         <widget class="QLabel" name="trick_level_minimal_logic_label">
-          <property name="text">
-           <string>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;/p&gt;&lt;p&gt;Minimal Logic is a setting that checks for the bare minimum items to create an almost pure random layout. This setting assumes the player has extensive knowledge of the game and will likely require Out of Bounds to complete.&lt;/p&gt;&lt;p&gt;There are no guarantees that a seed will be possible in this case.&lt;/p&gt;&lt;p&gt;{game_specific_text}&lt;/body&gt;&lt;/html&gt;</string>
-          </property>
-          <property name="alignment">
-           <set>Qt::AlignLeading|Qt::AlignLeft|Qt::AlignTop</set>
-          </property>
-          <property name="wordWrap">
-           <bool>true</bool>
-          </property>
-         </widget>
-        </item>
-        <item>
          <widget class="QCheckBox" name="underwater_abuse_check">
           <property name="text">
            <string>Allow Dangerous Gravity Suit Logic</string>
@@ -169,7 +107,7 @@
          </widget>
         </item>
         <item>
-         <widget class="Line" name="trick_level_line_3">
+         <widget class="Line" name="trick_level_line_2">
           <property name="orientation">
            <enum>Qt::Horizontal</enum>
           </property>
